# src/Parbat.Core/Parbat.Core.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parbat.Core.API",
    "description": "An open source ERP for Universities",
    "contact": {
      "name": "Dr. Muhammad Qasim Pasta",
      "url": "https://twitter.com/mqpasta",
      "email": "qasim.pasta@ieee.org"
    },
    "license": {
      "name": "The MIT License (MIT)",
      "url": "http://opensource.org/licenses/mit-license.php"
    },
    "version": "v2.0"
  },
  "paths": {
...
    "/api/v2.0/Component/Get/{id}": {
      "get": {
        "tags": [
          "Component"
        ],
        "summary": "Get a record",
        "operationId": "ComponentGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/Component/Update": {
      "put": {
        "tags": [
          "Component"
        ],
        "summary": "Update a component",
        "operationId": "ComponentUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Component/Create": {
      "post": {
        "tags": [
          "Component"
        ],
        "summary": "Create",
        "operationId": "ComponentCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Component/Delete/{id}": {
      "delete": {
        "tags": [
          "Component"
        ],
        "summary": "Delete",
        "operationId": "ComponentDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v2.0/Course/Get/{id}": {
      "get": {
        "tags": [
          "Course"
        ],
        "summary": "Get a record",
        "operationId": "CourseGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/Course/Update": {
      "put": {
        "tags": [
          "Course"
        ],
        "summary": "Update a course",
        "operationId": "CourseUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Course/Create": {
      "post": {
        "tags": [
          "Course"
        ],
        "summary": "Create",
        "operationId": "CourseCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Course/Delete/{id}": {
      "delete": {
        "tags": [
          "Course"
        ],
        "summary": "Delete",
        "operationId": "CourseDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v2.0/Course/ListCourseOrgUnit/{id}": {
      "get": {
        "tags": [
          "Course"
        ],
        "summary": "ListCourseOrgUnit",
        "operationId": "CourseListCourseOrgUnitById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Component": {
        "required": [
          "ComponentTypeID",
          "CurriculumID",
          "IsLimitCourses",
          "IsLimitCreditHrs",
          "Name",
          "ShortName",
          "TotalCourses",
          "TotalCredit"
        ],
        "type": "object",
        "properties": {
          "ComponentID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "ShortName": {
            "minLength": 1,
            "type": "string"
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "CurriculumID": {
            "type": "integer",
            "format": "int32"
          },
          "TotalCredit": {
            "type": "integer",
            "format": "int32"
          },
          "TotalCourses": {
            "type": "integer",
            "format": "int32"
          },
          "IsLimitCourses": {
            "type": "boolean"
          },
          "IsLimitCreditHrs": {
            "type": "boolean"
          },
          "ComponentTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Course": {
        "required": [
          "CourseTypeID",
          "LabCredit",
          "Name",
          "OwnerID",
          "ShortName",
          "TheoryCredit"
        ],
        "type": "object",
        "properties": {
          "CourseID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "ShortName": {
            "minLength": 1,
            "type": "string"
          },
          "OwnerID": {
            "type": "integer",
            "format": "int64"
          },
          "TheoryCredit": {
            "type": "integer",
            "format": "int32"
          },
          "LabCredit": {
            "type": "integer",
            "format": "int32"
          },
          "Code": {
            "type": "string",
            "nullable": true
          },
          "CourseTypeID": {
            "type": "integer",
            "format": "int64"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}